requirements.txt
boto==2.38.0
csscompressor==0.9.3
cssselect==0.9.1
django-appconf==1.0.1
django-countries==3.3
django-imagekit==3.2.6
django-phonenumber-field==0.7.2
django-pipeline==1.5.0
django-redis==4.1.0
django-storages==1.1.8
djangorestframework==3.1.2
djangorestframework-jsonp==1.0.0
docutils==0.7
futures==2.2.0
gunicorn==19.3.0
jmespath==0.7.1
jsmin==2.1.1
msgpack-python==0.4.6
netifaces==0.10.4
numpy==1.5.0
oauthlib==0.7.2
phonenumbers==7.0.5
pilkit==1.1.12
Pillow==2.8.1
ply==3.6
psycopg2
pycparser==2.10
pycrypto==2.3
python-memcached==1.54
queuelib==1.2.2
redis==2.10.3
reportlab==2.4
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
slimit==0.8.1
sorl-thumbnail==12.2
twython==3.2.0
